backend/requirements.txt
-i https://pypi.org/simple
aiohttp==3.9.5; python_version >= '3.8'
aiosignal==1.3.1; python_version >= '3.7'
annotated-types==0.6.0; python_version >= '3.8'
anyio==4.3.0; python_version >= '3.8'
async-generator==1.10; python_version >= '3.5'
async-timeout==4.0.3; python_version < '3.12.0'
asyncpg==0.29.0; python_full_version >= '3.8.0'
attrs==23.2.0; python_version >= '3.7'
backoff==2.2.1; python_version >= '3.7' and python_version < '4.0'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
black==24.4.2; python_version >= '3.8'
boto3==1.34.108; python_version >= '3.8'
botocore==1.34.108; python_version >= '3.8'
certifi==2024.2.2; python_version >= '3.6'
charset-normalizer==3.3.2; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coloredlogs==15.0.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
colorlog==6.8.2; python_version >= '3.6'
cssselect==1.2.0; python_version >= '3.7'
dataclasses-json==0.6.6; python_version >= '3.7' and python_version < '4.0'
diskcache==5.6.3; python_version >= '3'
distro==1.9.0; python_version >= '3.6'
dnspython==2.6.1; python_version >= '3.8'
duckduckgo-search==6.1.0; python_version >= '3.8'
ecdsa==0.19.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
email-validator==2.1.1; python_version >= '3.8'
fastapi==0.111.0; python_version >= '3.8'
fastapi-cli==0.0.3; python_version >= '3.8'
feedfinder2==0.0.4
feedparser==6.0.11; python_version >= '3.6'
filelock==3.14.0; python_version >= '3.8'
flake8==7.0.0; python_full_version >= '3.8.1'
flake8-black==0.3.6; python_version >= '3.7'
flashrank==0.2.5; python_version >= '3.6'
flatbuffers==24.3.25
frozenlist==1.4.1; python_version >= '3.8'
fsspec==2024.5.0; python_version >= '3.8'
gitdb==4.0.11; python_version >= '3.7'
gitpython==3.1.43; python_version >= '3.7'
googlemaps==4.10.0; python_version >= '3.5'
greenlet==3.0.3; python_version >= '3.7'
h11==0.14.0; python_version >= '3.7'
html5lib==1.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
httpcore==1.0.5; python_version >= '3.8'
httptools==0.6.1
httpx==0.27.0; python_version >= '3.8'
huggingface-hub==0.23.0; python_full_version >= '3.8.0'
humanfriendly==10.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
idna==3.7; python_version >= '3.5'
iniconfig==2.0.0; python_version >= '3.7'
jieba3k==0.35.1
jinja2==3.1.4; python_version >= '3.7'
jmespath==1.0.1; python_version >= '3.7'
joblib==1.4.2; python_version >= '3.8'
jq==1.7.0; python_version >= '3.5'
jsonpatch==1.33; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
jsonpointer==2.4; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
langchain==0.2.0; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-community==0.2.0; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-core==0.2.0; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-openai==0.1.7; python_full_version >= '3.8.1' and python_version < '4.0'
langchain-text-splitters==0.2.0; python_full_version >= '3.8.1' and python_version < '4.0'
langfuse==2.32.0; python_full_version >= '3.8.1' and python_version < '4.0'
langsmith==0.1.59; python_full_version >= '3.8.1' and python_version < '4.0'
llama-cpp-python==0.2.67; python_version >= '3.8'
lxml==5.2.2; python_version >= '3.6'
markdown-it-py==3.0.0; python_version >= '3.8'
markupsafe==2.1.5; python_version >= '3.7'
marshmallow==3.21.2; python_version >= '3.8'
mccabe==0.7.0; python_version >= '3.6'
mdurl==0.1.2; python_version >= '3.7'
mpmath==1.3.0
multidict==6.0.5; python_version >= '3.7'
mypy-extensions==1.0.0; python_version >= '3.5'
nest-asyncio==1.6.0; python_version >= '3.5'
newspaper3k==0.2.8
nltk==3.8.1; python_version >= '3.7'
nodeenv==1.8.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
numpy==1.26.4; python_version >= '3.9'
onnxruntime==1.18.0
openai==1.30.1; python_full_version >= '3.7.1'
orjson==3.10.3; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pathspec==0.12.1; python_version >= '3.8'
pillow==10.3.0; python_version >= '3.8'
platformdirs==4.2.2; python_version >= '3.8'
playwright==1.44.0; python_version >= '3.8'
pluggy==1.5.0; python_version >= '3.8'
protobuf==5.26.1; python_version >= '3.8'
psycopg2==2.9.9; python_version >= '3.7'
psycopg2-binary==2.9.9; python_version >= '3.7'
pyasn1==0.6.0; python_version >= '3.8'
pycodestyle==2.11.1; python_version >= '3.8'
pydantic==2.7.1; python_version >= '3.8'
pydantic-core==2.18.2; python_version >= '3.8'
pydantic-settings==2.2.1; python_version >= '3.8'
pyee==11.1.0; python_version >= '3.8'
pyflakes==3.2.0; python_version >= '3.8'
pygments==2.18.0; python_version >= '3.8'
pyinstrument==4.6.2; python_version >= '3.7'
pyreqwest-impersonate==0.4.5; python_version >= '3.8'
pyright==1.1.363; python_version >= '3.7'
pytest==8.2.0; python_version >= '3.8'
pytest-dotenv==0.5.2
pytest-mock==3.14.0; python_version >= '3.8'
python-dateutil==2.9.0.post0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
python-dotenv==1.0.1; python_version >= '3.8'
python-jose==3.3.0
python-multipart==0.0.9; python_version >= '3.8'
pyyaml==6.0.1; python_version >= '3.6'
regex==2024.5.15; python_version >= '3.8'
requests==2.31.0; python_version >= '3.7'
requests-file==2.0.0
rich==13.7.1; python_full_version >= '3.7.0'
rsa==4.9; python_version >= '3.6' and python_version < '4'
s3transfer==0.10.1; python_version >= '3.8'
setuptools==69.5.1; python_version >= '3.8'
sgmllib3k==1.0.0
shellingham==1.5.4; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2'
smmap==5.0.1; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.5; python_version >= '3.8'
sqlalchemy==2.0.30; python_version >= '3.7'
starlette==0.37.2; python_version >= '3.8'
sympy==1.12; python_version >= '3.8'
tenacity==8.3.0; python_version >= '3.8'
tiktoken==0.7.0; python_version >= '3.8'
tinysegmenter==0.3
tldextract==5.1.2; python_version >= '3.8'
tokenizers==0.19.1; python_version >= '3.7'
tqdm==4.66.4; python_version >= '3.7'
typer==0.12.3; python_version >= '3.7'
typing-extensions==4.11.0; python_version >= '3.8'
typing-inspect==0.9.0
ujson==5.10.0; python_version >= '3.8'
unidecode==1.3.8; python_version >= '3.5'
urllib3==2.2.1; python_version >= '3.10'
uvicorn[standard]==0.29.0; python_version >= '3.8'
uvloop==0.19.0
watchfiles==0.21.0
webencodings==0.5.1
websockets==12.0
wrapt==1.16.0; python_version >= '3.6'
yarl==1.9.4; python_version >= '3.7'
